[feat] add auth router and fix endpoint
