+++ backend/app/api/api.py
@@ -1,12 +1,19 @@
 from fastapi.responses import HTMLResponse
 from fastapi import APIRouter
 from app.api.routers.user import user_router
+from app.api.routers.auth import auth_router
 
 api_router = APIRouter(prefix="/api")
 
 @api_router.get("/")
 async def get():
     return HTMLResponse(html)
+
+api_router.include_router(
+    auth_router,
+    prefix="/auth",
+    tags=["auth"],
+)
 
 api_router.include_router(
     user_router,
